Move qcut draft
