+++ pennylane/transforms/__init__.py
@@ -74,48 +74,6 @@ There are also utility functions that take a circuit and return a DAG.
     ~transforms.CommutationDAG
     ~transforms.CommutationDAGNode
 
-Transform for circuit cutting
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-The :func:`~.cut_circuit` transform accepts a QNode and returns a new function that cuts the original circuit,
-allowing larger circuits to be split into smaller circuits that are compatible with devices that
-have a restricted number of qubits.
-
-.. autosummary::
-    :toctree: api
-
-    ~cut_circuit
-
-The :func:`~.cut_circuit_mc` transform is designed to be used for cutting circuits which contain :func:`~.sample`
-measurements and is implemented using a Monte Carlo method. Similarly to the :func:`~.cut_circuit`
-transform, this transform accepts a QNode and returns a new function that cuts the original circuit.
-This transform can also accept an optional classical processing function to calculate an
-expectation value.
-
-.. autosummary::
-    :toctree: api
-
-    ~cut_circuit_mc
-
-There are also low-level functions that can be used to build up the circuit cutting functionalities:
-
-.. autosummary::
-    :toctree: api
-
-    ~transforms.qcut.tape_to_graph
-    ~transforms.qcut.replace_wire_cut_nodes
-    ~transforms.qcut.fragment_graph
-    ~transforms.qcut.graph_to_tape
-    ~transforms.qcut.expand_fragment_tape
-    ~transforms.qcut.expand_fragment_tapes_mc
-    ~transforms.qcut.qcut_processing_fn
-    ~transforms.qcut.qcut_processing_fn_sample
-    ~transforms.qcut.qcut_processing_fn_mc
-    ~transforms.qcut.CutStrategy
-    ~transforms.qcut.kahypar_cut
-    ~transforms.qcut.place_wire_cuts
-    ~transforms.qcut.find_and_place_cuts
-
 Transforms for error mitigation
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
@@ -283,7 +241,5 @@ from .tape_expand import (
     set_decomposition,
 )
 from .transpile import transpile
-from . import qcut
-from .qcut import cut_circuit, cut_circuit_mc
 from .zx import to_zx, from_zx
 from .broadcast_expand import broadcast_expand
