+++ tests/transforms/test_qcut.py
@@ -33,9 +33,8 @@ from networkx import __version__ as networkx_version
 from scipy.stats import unitary_group
 
 import pennylane as qml
-from pennylane import numpy as np
+from pennylane import numpy as np, qcut
 from pennylane.queuing import WrappedObj
-from pennylane.transforms import qcut
 from pennylane.wires import Wires
 
 pytestmark = pytest.mark.qcut
@@ -3061,7 +3060,7 @@ class TestCutCircuitMCTransform:
             qml.CNOT(wires=[0, 1])
             return qml.sample(wires=[0, 1])
 
-        spy = mocker.spy(qcut.montecarlo, "qcut_processing_fn_mc")
+        spy = mocker.spy(pennylane.qcut.montecarlo, "qcut_processing_fn_mc")
         x = np.array(0.531, requires_grad=True)
         res = circuit(x)
 
@@ -3104,7 +3103,7 @@ class TestCutCircuitMCTransform:
             qml.CNOT(wires=[0, 1])
             return qml.sample(wires=[0, 1])
 
-        spy = mocker.spy(qcut.montecarlo, "qcut_processing_fn_mc")
+        spy = mocker.spy(pennylane.qcut.montecarlo, "qcut_processing_fn_mc")
 
         x = 0.4
         res = circuit(x)
@@ -3163,7 +3162,7 @@ class TestCutCircuitMCTransform:
 
             return qml.sample(wires=[1, 2, 3])
 
-        spy = mocker.spy(qcut.montecarlo, "qcut_processing_fn_mc")
+        spy = mocker.spy(pennylane.qcut.montecarlo, "qcut_processing_fn_mc")
 
         params = np.array([0.4, 0.5, 0.6, 0.7, 0.8], requires_grad=True)
         res = circuit(params)
@@ -3409,7 +3408,7 @@ class TestQCutProcessingFn:
             return qml.math.reshape(r, (4,) * (num_prep + num_meas))
 
         with monkeypatch.context() as m:
-            m.setattr(qcut.processing, "_process_tensor", mock_process_tensor)
+            m.setattr(pennylane.qcut.processing, "_process_tensor", mock_process_tensor)
             tensors_out = qcut._to_tensors(results, prepare_nodes, measure_nodes)
 
         for t1, t2 in zip(tensors, tensors_out):
@@ -3851,7 +3850,7 @@ class TestCutCircuitTransform:
             qml.RZ(0.133, wires=1)
             return qml.expval(qml.PauliZ(wires=[0]))
 
-        spy = mocker.spy(qcut.cutcircuit, "qcut_processing_fn")
+        spy = mocker.spy(pennylane.qcut.cutcircuit, "qcut_processing_fn")
         x = np.array(0.531, requires_grad=True)
         cut_circuit = qcut.cut_circuit(circuit, use_opt_einsum=use_opt_einsum)
 
@@ -3997,7 +3996,7 @@ class TestCutCircuitTransform:
             qml.CNOT(wires=[0, 1])
             return qml.expval(qml.PauliZ(0) @ qml.PauliZ(1))
 
-        spy = mocker.spy(qcut.cutcircuit, "qcut_processing_fn")
+        spy = mocker.spy(pennylane.qcut.cutcircuit, "qcut_processing_fn")
         x = np.array(0.531, requires_grad=True)
         cut_circuit = qcut.cut_circuit(circuit, use_opt_einsum=use_opt_einsum)
 
@@ -4045,7 +4044,7 @@ class TestCutCircuitTransform:
         )
 
         # Run once with original value
-        spy = mocker.spy(qcut.processing, "qcut_processing_fn")
+        spy = mocker.spy(pennylane.qcut.processing, "qcut_processing_fn")
         res = cut_circuit_trace(x)
 
         spy.assert_not_called()
@@ -4114,7 +4113,7 @@ class TestCutCircuitTransform:
         )
 
         # Run once with original value
-        spy = mocker.spy(qcut.cutcircuit, "qcut_processing_fn")
+        spy = mocker.spy(pennylane.qcut.cutcircuit, "qcut_processing_fn")
 
         # Note we call the function twice but assert qcut_processing_fn is called once. We expect
         # qcut_processing_fn to be called once during JIT compilation, with subsequent calls to
@@ -4186,7 +4185,7 @@ class TestCutCircuitTransform:
         cut_circuit_jit = jax.jit(qcut.cut_circuit(circuit, use_opt_einsum=use_opt_einsum))
 
         # Run once with original value
-        spy = mocker.spy(qcut.cutcircuit, "qcut_processing_fn")
+        spy = mocker.spy(pennylane.qcut.cutcircuit, "qcut_processing_fn")
 
         # Note we call the function twice but assert qcut_processing_fn is called once. We expect
         # qcut_processing_fn to be called once during JIT compilation, with subsequent calls to
@@ -4298,7 +4297,7 @@ class TestCutCircuitTransform:
             qml.CNOT(wires=[0, 1])
             return qml.expval(qml.PauliZ(wires=[0]))
 
-        spy = mocker.spy(qcut.processing, "contract_tensors")
+        spy = mocker.spy(pennylane.qcut.processing, "contract_tensors")
 
         x = 0.4
         res = circuit(x)
@@ -4372,7 +4371,7 @@ class TestCutCircuitTransform:
         res_expected = circuit(params)
         grad_expected = qml.grad(circuit)(params)
 
-        spy = mocker.spy(qcut.cutcircuit, "qcut_processing_fn")
+        spy = mocker.spy(pennylane.qcut.cutcircuit, "qcut_processing_fn")
         res = cut_circuit(params)
         spy.assert_called_once()
         grad = qml.grad(cut_circuit)(params)
@@ -4420,7 +4419,7 @@ class TestCutCircuitTransform:
 
         res_expected = circuit()
 
-        spy = mocker.spy(qcut.cutcircuit, "qcut_processing_fn")
+        spy = mocker.spy(pennylane.qcut.cutcircuit, "qcut_processing_fn")
         res = cut_circuit()
         spy.assert_called_once()
 
@@ -4513,8 +4512,8 @@ class TestCutCircuitExpansion:
             qml.RY(0.4, wires=0)
             return qml.apply(measurement)
 
-        spy = mocker.spy(qcut.cutcircuit, "_qcut_expand_fn")
-        spy_mc = mocker.spy(qcut.montecarlo, "_qcut_expand_fn")
+        spy = mocker.spy(pennylane.qcut.cutcircuit, "_qcut_expand_fn")
+        spy_mc = mocker.spy(pennylane.qcut.montecarlo, "_qcut_expand_fn")
 
         kwargs = {"shots": 10} if measurement.return_type is qml.measurements.Sample else {}
         cut_transform(circuit, device_wires=[0])(**kwargs)
@@ -4533,9 +4532,9 @@ class TestCutCircuitExpansion:
             qml.apply(measurement)
 
         tape = qml.tape.QuantumScript.from_queue(q)
-        spy = mocker.spy(qcut.tapes, "_qcut_expand_fn")
-        spy_cc = mocker.spy(qcut.cutcircuit, "_qcut_expand_fn")
-        spy_mc = mocker.spy(qcut.montecarlo, "_qcut_expand_fn")
+        spy = mocker.spy(pennylane.qcut.tapes, "_qcut_expand_fn")
+        spy_cc = mocker.spy(pennylane.qcut.cutcircuit, "_qcut_expand_fn")
+        spy_mc = mocker.spy(pennylane.qcut.montecarlo, "_qcut_expand_fn")
 
         kwargs = {"shots": 10} if measurement.return_type is qml.measurements.Sample else {}
         cut_transform(tape, device_wires=[0], **kwargs)
@@ -4583,8 +4582,8 @@ class TestCutCircuitExpansion:
         qnode = qml.QNode(circuit, dev_big)
         qnode_cut = qcut.cut_circuit(qml.QNode(circuit, dev_cut))
 
-        spy_tapes = mocker.spy(qcut.tapes, "_qcut_expand_fn")
-        spy_cc = mocker.spy(qcut.cutcircuit, "_qcut_expand_fn")
+        spy_tapes = mocker.spy(pennylane.qcut.tapes, "_qcut_expand_fn")
+        spy_cc = mocker.spy(pennylane.qcut.cutcircuit, "_qcut_expand_fn")
 
         res = qnode_cut(template_weights)
 
@@ -4616,8 +4615,8 @@ class TestCutCircuitExpansion:
 
         qnode_cut = qcut.cut_circuit_mc(qml.QNode(circuit, dev_cut))
 
-        spy_tapes = mocker.spy(qcut.tapes, "_qcut_expand_fn")
-        spy_mc = mocker.spy(qcut.montecarlo, "_qcut_expand_fn")
+        spy_tapes = mocker.spy(pennylane.qcut.tapes, "_qcut_expand_fn")
+        spy_mc = mocker.spy(pennylane.qcut.montecarlo, "_qcut_expand_fn")
 
         qnode_cut(template_weights)
 
@@ -5224,7 +5223,7 @@ class TestAutoCutCircuit:
         res_expected = circuit(params)
         grad_expected = qml.grad(circuit)(params)
 
-        spy = mocker.spy(qcut.cutcircuit, "qcut_processing_fn")
+        spy = mocker.spy(pennylane.qcut.cutcircuit, "qcut_processing_fn")
         res = cut_circuit(params)
         spy.assert_called_once()
         grad = qml.grad(cut_circuit)(params)
@@ -5267,7 +5266,7 @@ class TestAutoCutCircuit:
 
         res_expected = circuit()
 
-        spy = mocker.spy(qcut.cutcircuit, "qcut_processing_fn")
+        spy = mocker.spy(pennylane.qcut.cutcircuit, "qcut_processing_fn")
         res = cut_circuit()
         spy.assert_called_once()
 
@@ -5419,7 +5418,7 @@ class TestRedirect:
     """Tests that redirect in qcut.__init__ works to maintain import pathways while reorganizing files"""
 
     def test_qcut_redirects_to_qcut_qcut(self):
-        assert qml.transforms.qcut._prep_one_state == qml.transforms.qcut.qcut._prep_one_state
+        assert qml.transforms.qcut._prep_one_state == pennylane.qcut.qcut._prep_one_state
 
 
 class TestCutCircuitWithHamiltonians:
@@ -5486,7 +5485,7 @@ class TestCutCircuitWithHamiltonians:
         res_expected = circuit(params)
         grad_expected = qml.grad(circuit)(params)
 
-        spy = mocker.spy(qcut.cutcircuit, "qcut_processing_fn")
+        spy = mocker.spy(pennylane.qcut.cutcircuit, "qcut_processing_fn")
         res = cut_circuit(params)
         assert spy.call_count == len(hamiltonian.ops)
 
@@ -5539,7 +5538,7 @@ class TestCutCircuitWithHamiltonians:
 
         res_expected = circuit()
 
-        spy = mocker.spy(qcut.cutcircuit, "qcut_processing_fn")
+        spy = mocker.spy(pennylane.qcut.cutcircuit, "qcut_processing_fn")
         res = cut_circuit()
         assert spy.call_count == len(hamiltonian.ops)
         assert np.isclose(res, res_expected, atol=1e-8)
